+++ src/main/java/ar/edu/utn/frba/dds/simeal/models/entities/colaboraciones/Colaboracion.java
@@ -5,6 +5,7 @@ import ar.edu.utn.frba.dds.simeal.models.entities.personas.Colaborador;
 
 
 public interface Colaboracion {
-  Float calcularReconocimientoParcial();
+  double calcularReconocimientoParcial();
   Colaborador getColaborador();
+
 }
