+++ src/main/java/ar/edu/utn/frba/dds/simeal/models/entities/colaboraciones/oferta/Oferta.java
@@ -29,8 +29,8 @@ public class Oferta implements Colaboracion {
 */
 
   @Override
-  public Float calcularReconocimientoParcial() {
-    return 0F;
+  public double calcularReconocimientoParcial() {
+    return 0;
   }
 
   public void canjear() {
